+++ core/src/main/scala-2/slinky/core/annotations/react.scala
