+++ core/src/main/scala/slinky/core/BaseComponentWrapper.scala
@@ -6,9 +6,6 @@ import slinky.readwrite.{Reader, Writer}
 import scala.scalajs.js
 import scala.scalajs.js.ConstructorTag
 
-import scala.language.experimental.macros
-import scala.reflect.macros.whitebox
-
 final class KeyAndRefAddingStage[D](private val args: js.Array[js.Any]) extends AnyVal {
   @inline def withKey(key: String): KeyAndRefAddingStage[D] = {
     if (args(0) == null) {
@@ -308,38 +305,4 @@ object BaseComponentWrapper {
       middleware(orig(constructor, componentObject), componentObject)
     }
   }
-}
-
-trait StateReaderProvider extends js.Object
-object StateReaderProvider {
-  def impl(c: whitebox.Context): c.Expr[StateReaderProvider] = {
-    import c.universe._
-    val compName = c.internal.enclosingOwner.owner.asClass
-    val q"$_; val x: $typedReaderType = null" = c.typecheck(
-      q"@_root_.scala.annotation.unchecked.uncheckedStable val comp: $compName = null; val x: _root_.slinky.readwrite.Reader[comp.State] = null"
-    ) // scalafix:ok
-    val tpcls = c.inferImplicitValue(typedReaderType.tpe.asInstanceOf[c.Type], silent = false)
-    c.Expr(
-      q"if (_root_.scala.scalajs.LinkingInfo.productionMode) null else $tpcls.asInstanceOf[_root_.slinky.core.StateReaderProvider]"
-    )
-  }
-
-  implicit def get: StateReaderProvider = macro impl
-}
-
-trait StateWriterProvider extends js.Object
-object StateWriterProvider {
-  def impl(c: whitebox.Context): c.Expr[StateWriterProvider] = {
-    import c.universe._
-    val compName = c.internal.enclosingOwner.owner.asClass
-    val q"$_; val x: $typedReaderType = null" = c.typecheck(
-      q"@_root_.scala.annotation.unchecked.uncheckedStable val comp: $compName = null; val x: _root_.slinky.readwrite.Writer[comp.State] = null"
-    ) // scalafix:ok
-    val tpcls = c.inferImplicitValue(typedReaderType.tpe.asInstanceOf[c.Type], silent = false)
-    c.Expr(
-      q"if (_root_.scala.scalajs.LinkingInfo.productionMode) null else $tpcls.asInstanceOf[_root_.slinky.core.StateWriterProvider]"
-    )
-  }
-
-  implicit def get: StateWriterProvider = macro impl
 }
