+++ core/src/main/scala/slinky/core/ExternalComponent.scala
@@ -5,8 +5,6 @@ import slinky.readwrite.Writer
 
 import scala.scalajs.js
 import scala.scalajs.js.|
-import scala.language.experimental.macros
-import scala.reflect.macros.whitebox
 
 final class BuildingComponent[E, R <: js.Object](private val args: js.Array[js.Any]) extends AnyVal {
   def apply(mods: TagMod[E]*): BuildingComponent[E, R] = {
@@ -116,19 +114,3 @@ abstract class ExternalComponentNoPropsWithRefType[R <: js.Object]
     extends ExternalComponentNoPropsWithAttributesWithRefType[Nothing, R]
 
 abstract class ExternalComponentNoProps extends ExternalComponentNoPropsWithAttributes[Nothing]
-
-// same as PropsWriterProvider except it always returns the typeclass instead of nulling it out in fullOpt mode
-trait ExternalPropsWriterProvider extends js.Object
-object ExternalPropsWriterProvider {
-  def impl(c: whitebox.Context): c.Expr[ExternalPropsWriterProvider] = {
-    import c.universe._
-    val compName = c.internal.enclosingOwner.owner.asClass
-    val q"$_; val x: $typedReaderType = null" = c.typecheck(
-      q"@_root_.scala.annotation.unchecked.uncheckedStable val comp: $compName = null; val x: _root_.slinky.readwrite.Writer[comp.Props] = null"
-    ) // scalafix:ok
-    val tpcls = c.inferImplicitValue(typedReaderType.tpe.asInstanceOf[c.Type])
-    c.Expr(q"$tpcls.asInstanceOf[_root_.slinky.core.ExternalPropsWriterProvider]")
-  }
-
-  implicit def get: ExternalPropsWriterProvider = macro impl
-}
